Refactor lookup table to be more efficient and track multi-chromosome
